requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
paddleocr>=2.7.0
paddlepaddle>=2.5.0
pdf2image>=1.16.0
Pillow>=10.0.0
python-docx>=1.1.0
httpx>=0.25.0
